deprecated/bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "1.78"
description = "Bot"
license = "MIT"
authors = ["elaresai <elaresai@gmail.com>"]
maintainers = ["elaresai <elaresai@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/elaresai/pudgeland"
repository = "https://github.com/elaresai/pudgeland"
classifiers = [
    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pre-commit = "^3.3.3"
ruff = "^0.0.282"
black = "^23.7.0"
python-dotenv = "^1.0.0"
attrs = "^23.1.0"
hikari = "^2.0.0.dev120"
hikari-crescent = "^0.6.2"
hikari-miru = "^3.1.2"
prisma = "^0.9.1"
msgspec = "^0.17.0"
alhaitham = "^0.1.3"
nahida = "^0.1.3"
tighnari = "^0.1.3"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
